* gmn/master: (266 commits)
  fixed case when no stars are selected and ECSV is saved
  added mov format
  Fix trackstack integration of constellations
  Fix command line, start integrate trackstack
  Remove unused imports
  Add DrawConstellations.py
  updated radiants with GMN-measured values
  added tau herculids to established list
  added script for fitting shower activity curves
  added line functions
  remove debug statement
  correct comment
  fix to trackstack bug #70 raised by D Rollinson
  added blocking uploads if the defualt station code is uped
  Fixed sorting
  Combine frame clippings that occured at same time into a single video
  fixed function for finding FTPdetectinfos
  added printing verbose status info
  added manually specifying binning
  fixed year of AND outburst
  ...

+++ RMS/Formats/AsgardEv.py
@@ -65,10 +65,10 @@ class AsgardEv(object):
         fr_arr, time_arr, sum_arr, seq_arr, cx_arr, cy_arr, th_arr, phi_arr, lsp_arr, mag_arr, flag_arr, \
             bak_arr, max_arr = self.data.T
 
-        fr_arr = fr_arr.astype(np.int)
-        sum_arr = sum_arr.astype(np.int)
-        seq_arr = seq_arr.astype(np.int)
-        flag_arr = flag_arr.astype(np.int)
+        fr_arr = fr_arr.astype(int)
+        sum_arr = sum_arr.astype(int)
+        seq_arr = seq_arr.astype(int)
+        flag_arr = flag_arr.astype(int)
 
         for i in range(len(self.data)):
             out_str += "{:5d} {:10.6f} {:6d} {:7d} {:11.6f} {:11.6f} {:10.6f} {:11.6f} {:7.3f} {:6.2f}  {:04d} {:5.1f} {:6.1f}\n".format(\
